The Windows ending cause my tests to fail :/
I just changed it to the "right" path separator. Maybe, if the
`path->string` part of the `git repo path` test is returning
system specific path separators at the end we should fix it
in the test functions instead of relying on a string literal.
I am willing to work on that.